clientes salvos

+++ src/main/java/org/wayster/com/emprestimos/Repository/EmprestimoRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface EmprestimoRepository extends JpaRepository<Emprestimo, Integer> {
+public interface EmprestimoRepository extends JpaRepository<Emprestimo, Long> {
 
     List<Emprestimo> findByDataVencimento(LocalDate dataVencimento);
